One fix in background_subtraction(), but ...
Did one fix in background_subtraction(): before, the sky fit was evaluated
(and returned) for all pixels in the 2D image, and now it's only done for
the pixels in the given slit.
The fit from background_subtraction() is still bad, and I've realised the
tilts are wrong: they tilt the wrong way!