The echelle mode tilts are now in the right direction (bug fixed)
Fixed bug in echelle_tilt() in artrace.py, following Ryan's advice,
and the tilts in echelle mode are now in the right direction (for the
X-shooter VIS data with large tilts this is very visible). However,
the exact fit of the tilts for the X-shooter VIS data is not good:
in some parts of the frame the fitted tilts are too steep, and in some
too shallow. This means that the sky fit made by background_subtraction()
is poor.
MAde a small change in reduce_echelle() in arproc.py related to the
naming of the Ginga frames (if that's the word) that I'm using to debug
the sky subtraction in reduce_echelle()